fix(menu): show args and valid code snippets in menu doc
